+++ exporters/variables-scss/src/content/generator.ts
@@ -1,6 +1,6 @@
 import { TokenGroup, Token, TokenType, DimensionToken } from '@supernovaio/sdk-exporters';
 import { CSSHelper, NamingHelper, StringCase } from '@supernovaio/export-helpers';
-import { generateSpacesObject } from './token';
+import { exportConfiguration } from '../index';
 
 type TokenHandler = (token: Token, mappedTokens: Map<string, Token>, tokenGroups: TokenGroup[]) => string | null;
 
@@ -15,10 +15,25 @@ const tokenVariableName = (token: Token, tokenGroups: Array<TokenGroup>, without
   return NamingHelper.codeSafeVariableNameForToken(token, StringCase.paramCase, parent, '');
 };
 
-export const generate = (tokens: Array<Token>, mappedTokens: Map<string, Token>, tokenGroups: Array<TokenGroup>) => {
-  const spacingContent = createSpacingContent(tokens, mappedTokens, tokenGroups);
+const addDisclaimer = (content: string): string => {
+  if (exportConfiguration.generateDisclaimer) {
+    return `/* This file was generated by Supernova, don't change by hand */\n${content}`;
+  }
 
-  return spacingContent;
+  return content;
+};
+
+export const generateContent = (
+  tokens: Array<Token>,
+  mappedTokens: Map<string, Token>,
+  tokenGroups: Array<TokenGroup>,
+) => {
+  return [
+    {
+      fileName: '_spacing.scss',
+      content: createSpacingContent(tokens, mappedTokens, tokenGroups),
+    },
+  ];
 };
 
 const filterTokensByType = (tokens: Token[], type: TokenType): Token[] => {
@@ -59,10 +74,37 @@ export const measuresTokenToCSS = (
   return `$${name}: ${value}${unit} !default;`;
 };
 
-const createSpacingContent = (tokens: Token[], mappedTokens: Map<string, Token>, tokenGroups: Array<TokenGroup>) => {
+export const createSpacingContent = (
+  tokens: Token[],
+  mappedTokens: Map<string, Token>,
+  tokenGroups: Array<TokenGroup>,
+) => {
   const filteredSpacingTokens = filterTokensByType(tokens, TokenType.dimension);
   const spacingTokensToCSS = tokensToCSS(filteredSpacingTokens, measuresTokenToCSS, mappedTokens, tokenGroups);
   const spacingObject = generateSpacesObject(tokens, mappedTokens, tokenGroups);
 
-  return `${spacingTokensToCSS}\n\n${spacingObject}`;
+  return addDisclaimer(`${spacingTokensToCSS}\n\n${spacingObject}`);
+};
+
+export const generateSpacesObject = (
+  tokens: Array<Token>,
+  mappedTokens: Map<string, Token>,
+  tokenGroups: Array<TokenGroup>,
+): string => {
+  let spacesObject = '$spaces: (\n';
+
+  tokens.forEach((token) => {
+    if (token.tokenType === TokenType.dimension && token.origin?.name?.includes('Spacing system')) {
+      const dimensionToken = token as DimensionToken;
+      const name = tokenVariableName(dimensionToken, tokenGroups, true);
+      const numericPart = name.match(/\d+/)?.[0];
+      if (numericPart) {
+        spacesObject += `    ${numericPart}: $${name},\n`;
+      }
+    }
+  });
+
+  spacesObject += ') !default;';
+
+  return spacesObject;
 };
