+++ exporters/variables-scss/src/index.ts
@@ -4,16 +4,10 @@ import {
   PulsarContext,
   RemoteVersionIdentifier,
   Supernova,
-  TokenType,
-  Token,
-  TokenGroup,
   OutputTextFile,
 } from '@supernovaio/sdk-exporters';
 import { ExporterConfiguration } from '../config';
-import { generateSpacesObject, measuresTokenToCSS, otherTokenToCSS } from './content/token';
-import { generate } from './content/generator';
-
-type TokenHandler = (token: Token, mappedTokens: Map<string, Token>, tokenGroups: TokenGroup[]) => string | null;
+import { generateContent } from './content/generator';
 
 // @ts-ignore-next-line
 Pulsar.export(async (sdk: Supernova, context: PulsarContext): Promise<Array<AnyOutputFile>> => {
@@ -48,26 +42,6 @@ Pulsar.export(async (sdk: Supernova, context: PulsarContext): Promise<Array<AnyO
   // Convert all color tokens to CSS variables
   const mappedTokens = new Map(tokens.map((token) => [token.id, token]));
 
-  const test = generate(tokens, mappedTokens, tokenGroups);
-
-  const addDisclaimer = (content: string): string => {
-    if (exportConfiguration.generateDisclaimer) {
-      return `/* This file was generated by Supernova, don't change by hand */\n${content}`;
-    }
-
-    return content;
-  };
-
-  const createCssVariable = (type: TokenType, handler: TokenHandler): string => {
-    const result = tokens
-      .filter((t) => t.tokenType === type)
-      .map((token) => handler(token, mappedTokens, tokenGroups))
-      .filter(Boolean)
-      .join('\n');
-
-    return addDisclaimer(result);
-  };
-
   function createTextFile(relativePath: string, fileName: string, content: string): OutputTextFile {
     return FileHelper.createTextFile({
       relativePath,
@@ -76,14 +50,12 @@ Pulsar.export(async (sdk: Supernova, context: PulsarContext): Promise<Array<AnyO
     });
   }
 
-  const measuresContent = createCssVariable(TokenType.dimension, measuresTokenToCSS);
-  const spacesContent = generateSpacesObject(tokens, mappedTokens, tokenGroups);
-  const combinedContent = `${measuresContent}\n\n${spacesContent}`;
-  const otherContent = createCssVariable(TokenType.dimension, otherTokenToCSS);
+  const contents = generateContent(tokens, mappedTokens, tokenGroups);
 
   return [
-    createTextFile('./', '_measures.scss', combinedContent),
-    createTextFile('./', '_other.scss', otherContent),
+    ...contents.map((c) => {
+      return createTextFile('./', c.fileName, c.content);
+    }),
     createTextFile(
       './original-data/',
       '_original.json',
@@ -100,7 +72,6 @@ Pulsar.export(async (sdk: Supernova, context: PulsarContext): Promise<Array<AnyO
         2,
       ),
     ),
-    createTextFile('./original-data/', '_test.scss', test),
   ];
 });
